feat: require secret to be managed by nauth on update (#52)
To ensure nauth does not mutate secrets not managed by the nauth
operator, we verify that the existing secret is labelled with
`nauth.io/managed: true`.

Closes #26 

By submitting a request, you represent that you have the right to
license
your contribution to the community, and agree that your contributions
are
licensed under the [License](LICENSE).