refactor: flatten package structure (#86)
The package structure was unnecessary complicated. This PR simplifies by
flattening the package structure. The method names for some structs have
been updated where necessary as part of the move to a package.

**New package structure:**
```
internal/
    account/         - Account related logic
    controller/      - Controllers (reconcilers)
    k8s/             - K8S clients
    nats/            - Nats client
    user/            - User related logic
```

By submitting a request, you represent that you have the right to
license
your contribution to the community, and agree that your contributions
are
licensed under the [License](LICENSE).